chore: implement randomly choosing word function on MoeDictionary
